+++ .gitignore
@@ -16,6 +16,6 @@ core.*
 aam/__pycache__/*
 aam/.ipynb_checkpoints/*
 aam/.DS_Store
-sepsis/data/*
-sepsis/.ipynb_checkpoints/*
-sepsis/__pycache__/*
+attention_regression/data/*
+attention_regression/.ipynb_checkpoints/*
+attention_regression/__pycache__/*
