+++ aam/cli_util.py
@@ -3,66 +3,84 @@ import click
 
 def aam_model_options(func):
     model_options = [
-        click.option('--batch-size',
-                     default=8,
-                     show_default=True,
-                     type=int),
-        click.option('--train-percent',
-                     default=0.8,
-                     show_default=True,
-                     type=float),
-        click.option('--epochs',
-                     default=100,
-                     show_default=True,
-                     type=int),
-        click.option('--repeat',
-                     default=1,
-                     show_default=True,
-                     type=int),
-        click.option('--dropout',
-                     default=0.8,
-                     show_default=True,
-                     type=float),
-        click.option('--pca-hidden-dim',
-                     default=256,
-                     show_default=True,
-                     type=int),
-        click.option('--pca-heads',
-                     default=64,
-                     show_default=True,
-                     type=int),
-        click.option('--pca-layers',
-                     default=6,
-                     show_default=True,
-                     type=int),
-        click.option('--dff',
-                     default=2048,
-                     show_default=True,
-                     type=int),
-        click.option('--d-model',
-                     default=128,
-                     show_default=True,
-                     type=int),
-        click.option('--enc_layers',
-                     default=6,
-                     show_default=True,
-                     type=int),
-        click.option('--enc-heads',
-                     default=8,
-                     show_default=True,
-                     type=int),
-        click.option('--output-dim',
-                     default=1,
-                     show_default=True,
-                     type=int),
-        click.option('--lr',
-                     default=0.0005,
-                     show_default=True,
-                     type=float),
-        click.option('--max-bp',
-                     default=100,
-                     show_default=True,
-                     type=int),
+        click.option(
+            '--p-batch-size',
+            default=32,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-epochs',
+            default=1000,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-repeat',
+            default=5,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-dropout',
+            default=0.1,
+            show_default=True,
+            type=float
+        ),
+        click.option(
+            '--p-token-dim',
+            default=512,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-features-to-add-rate',
+            default=1.,
+            show_default=True,
+            type=float
+        ),
+        click.option(
+            '--p-ff-d-model',
+            default=128,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-ff-clr',
+            default=1024,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-pca-heads',
+            default=8,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-enc-layers',
+            default=2,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-enc-heads',
+            default=8,
+            show_default=True,
+            type=int
+        ),
+        click.option(
+            '--p-lr',
+            default=0.001,
+            show_default=True,
+            type=float
+        ),
+        click.option(
+            '--p-report-back-after',
+            default=5,
+            show_default=True,
+            type=int
+        ),
     ]
 
     for option in reversed(model_options):
